clean-up Order component
